Big changes to the Project Analysis.
I completely remade the User Interface Sketches section to better reflect what the application is about. I made some mistakes by making it 'meetup oriented', which misses the point of CineMatch, where it is originally about meeting just one other person. These new wireframes form the start of a (high-fidelity) prototype I'm working on that will be subject to usability testing soon.In addition, I rewrote US-04 to allow users to sign up with just SSO because setting up a new login/sign up system is out of scope for this project. It would be too time-consuming. It also enhances user experience, so I'm satisfied with this decision.